Add new testcases for nats sink
